backend/requirements.txt
flask
flask_cors
pika
pymongo
psycopg2-binary
minio
sentence-transformers
transformers
PyMuPDF
python-dotenv